Create 800-ALPHANUM
The numbers 2 - 9 (inclusive) on a telephone keypad have letters associated with them. 
Companies often select 800 telephone numbers because the digits of the number can be 
remembered by a word or name. For example:

The number:   1.800.468.6455

Can be remembered as:  1.800.GOT.MILK

You are to write a program that will print to the screen all of the words (and the associated numbers) 
that could be generated through an 800 phone number that has the prefix 555.  Remember that the 
numbers 1 and 0 on a keypad do not have letters associated with them.

Here are some examples of the words that would be generated and the format for your output: 
Number 	 Word
1.800.555.2222 	 1.800.555.AAAA
1.800.555.2222 	1.800.555.AAAB
1.800.555.2222 	1.800.555.AAAC
1.800.555.2223 	1.800.555.AAAD
1.800.555.2223 	 1.800.555.AAAE
1.800.555.2223 	 1.800.555.AAAF
1.800.555.2224 	 1.800.555.AAAG
1.800.555.2224 	 1.800.555.AAAH
1.800.555.2224 	 1.800.555.AAAI

Name your program PhoneNumbers.py

HINT:

 

I know that the assignment talks about phone numbers.  Don’t think in numbers but instead in letters, 
the letters possible in each of the last four digits of the phone number. Those letters can be contained 
in a single list of 26.

You are to generate the many (456,976) combinations of the letters, placing one letters in each of four 
different positions.  

Think of each position as a gear with 26 letters on its’ outer edge, each gear has all of the letters of 
the alphabet printed on it (one after another going around the outside edge of the gear), so you have four 
gears that are the same. 

Let’s assume that the gears are in the order (left to right) from most significant to least significant 
and are named gear4,  gear3, gear2, gear1. Each gear shows a single letter at a time. The gears all start 
with the letter ‘a’ exposed. After printing the exposed letters (aaaa) gear1 rotates one position.  The 
letters are now aaab. After printing the exposed letters (aaab) gear1 rotates one position.  The letters 
are now aaac.  After the 25th rotation the exposed letters are printed (aaaz). The next rotation (the 26th) 
returns gear1 to the ‘a’ position and causes gear2 to rotate a single position. The exposed letters are aaba, 
which are printed, gear1 then rotates a single position and the exposed letters are aabb.

As you can see the pattern is: with every complete (i.e. 26) rotation of a gear the previous gear rotates 
a single time.

Efficient, easy to read Python code is known as “Pythonic” code. Try to make your code as Pythonic as possible.